Main renamed to make more proper name. VictoryChecker added to check trhe vistory conditions. TODO: implement VictoryChecker
