+++ src/main/java/GameEntryPoint.java
@@ -1,0 +1,8 @@
+import tictactoegame.*;
+
+public class GameEntryPoint {
+    public static void main(String[] args) {
+        TicTacToeGame game = new TicTacToeGame();
+        game.play();
+    }
+}
