--- src/main/java/Main.java
@@ -1,8 +1,0 @@
-import game.Game;
-
-public class Main {
-    public static void main(String[] args) {
-        Game game = new Game();
-        game.play();
-    }
-}
