+++ src/main/java/board/VictoryChecker.java
@@ -1,7 +1,8 @@
 package board;
 
 public class VictoryChecker {
-   // public boolean checkVictory(PossibleSequences possibleSequences, Mark mark) {
+    //public boolean checkVictory(PossibleSequences possibleSequences, Mark mark) {
 
-   // }
+    //}
+
 }
